group routes for clients module
